move auditor tests to test/
